+++ db/migrate/20220203013803_add_folder_id_to_pages.rb
@@ -1,0 +1,6 @@
+class AddFolderIdToPages < ActiveRecord::Migration[6.1]
+  def change
+    add_column :pages, :folder_id, :bigint
+    add_index :pages, :folder_id
+  end
+end